refactor: Extract shared git exec environment and extractTitleFromDescription
