+++ apps/server/src/services/merge-eligibility-service.ts
@@ -10,41 +10,12 @@ import type { AutoMergeSettings, AutoMergeCheckType } from '@protolabsai/types';
 import { DEFAULT_AUTO_MERGE_SETTINGS } from '@protolabsai/types';
 import { exec } from 'child_process';
 import { promisify } from 'util';
+import { createGitExecEnv } from '@protolabsai/git-utils';
 
 const execAsync = promisify(exec);
 const logger = createLogger('MergeEligibility');
 
-// Extended PATH for finding gh CLI (same pattern as git-workflow-service)
-const pathSeparator = process.platform === 'win32' ? ';' : ':';
-const additionalPaths: string[] = [];
-
-if (process.platform === 'win32') {
-  if (process.env.LOCALAPPDATA) {
-    additionalPaths.push(`${process.env.LOCALAPPDATA}\\Programs\\Git\\cmd`);
-  }
-  if (process.env.PROGRAMFILES) {
-    additionalPaths.push(`${process.env.PROGRAMFILES}\\Git\\cmd`);
-  }
-  if (process.env['ProgramFiles(x86)']) {
-    additionalPaths.push(`${process.env['ProgramFiles(x86)']}\\Git\\cmd`);
-  }
-} else {
-  additionalPaths.push(
-    '/opt/homebrew/bin',
-    '/usr/local/bin',
-    '/home/linuxbrew/.linuxbrew/bin',
-    `${process.env.HOME}/.local/bin`
-  );
-}
-
-const extendedPath = [process.env.PATH, ...additionalPaths.filter(Boolean)]
-  .filter(Boolean)
-  .join(pathSeparator);
-
-const execEnv = {
-  ...process.env,
-  PATH: extendedPath,
-};
+const execEnv = createGitExecEnv();
 
 /**
  * PRCheckStatus - Status of a single PR check
